+++ packages/mdc-icon-toggle/index.js
@@ -21,9 +21,9 @@
  * THE SOFTWARE.
  */
 
-import MDCComponent from '@material/base/component';
+import MDCComponent from '@rsmdc/base/component';
 import MDCIconToggleFoundation from './foundation';
-import {MDCRipple, MDCRippleFoundation} from '@material/ripple/index';
+import {MDCRipple, MDCRippleFoundation} from '@rsmdc/ripple/index';
 
 /**
  * @extends {MDCComponent<!MDCIconToggleFoundation>}