+++ packages/mdc-menu/foundation.js
@@ -21,11 +21,11 @@
  * THE SOFTWARE.
  */
 
-import MDCFoundation from '@material/base/foundation';
+import MDCFoundation from '@rsmdc/base/foundation';
 import {MDCMenuAdapter} from './adapter';
 import {cssClasses, strings} from './constants';
-import {MDCMenuSurfaceFoundation} from '@material/menu-surface/foundation';
-import MDCListFoundation from '@material/list/foundation';
+import {MDCMenuSurfaceFoundation} from '@rsmdc/menu-surface/foundation';
+import MDCListFoundation from '@rsmdc/list/foundation';
 
 const ELEMENTS_KEY_ALLOWED_IN = ['input', 'button', 'textarea', 'select'];