+++ scripts/rewrite-sass-import-statements-for-closure.js
@@ -25,7 +25,7 @@
  * @fileoverview Rewrites import statements such that:
  *
  * ```js
- * import [<SPECIFIERS> from] '@material/$PKG[/files...]';
+ * import [<SPECIFIERS> from] '@rsmdc/$PKG[/files...]';
  * ```
  * becomes
  * ```js
@@ -80,7 +80,7 @@ function rewriteImportDeclaration(importSource, srcFile, rootDir) {
   const pathParts = importSource.split('/');
   const isMDCImport = pathParts[0] === '@material';
   if (isMDCImport) {
-    const modName = pathParts[1]; // @material/<modName>
+    const modName = pathParts[1]; // @rsmdc/<modName>
     const atMaterialReplacementPath = `${rootDir}/${modName}`;
     const rewrittenImportSource = [atMaterialReplacementPath].concat(pathParts.slice(2)).join('/');
     importSource = rewrittenImportSource;